+++ compiler/parse_tree_out_misc.m
@@ -281,6 +281,9 @@ goal_warning_to_string(Warning) = Str :-
         Warning = goal_warning_singleton_vars,
         Str = "singleton_vars"
     ;
+        Warning = goal_warning_repeated_singleton_vars,
+        Str = "repeated_singleton_vars"
+    ;
         Warning = goal_warning_suspicious_recursion,
         Str = "suspicious_recursion"
     ;